fix tests, move aave-credit test to integration tests.
